requirements.txt
feedparser==6.0.11
trafilatura==1.7.0
openai>=1.0.0
PyYAML>=6.0.1
typer>=0.12.0
rich>=13.7.0
python-dateutil>=2.9.0.post0
pytz>=2024.1
sqlite-utils>=3.37
tenacity>=8.5.0
tqdm>=4.66.0
gspread>=6.0.0
google-auth>=2.30.0
